Remove settors, optimize slider updating
gettors and settors in JS are kinda usless since everything is public
and it's not worth the effort of using closures to make things public,
If you want to write an unbreakable library, don't use JS.
For now we'll instead concentrat on making the public interface clear
is '_' notation for private things, (TODO). This commits removes
settors.

Also on slider change I was updating every slider on the page istead of
using the event and focusing on the slider that changed, now fixed.